requirements.txt
fastapi
uvicorn
speedtest-cli
httpx
psutil
python-dotenv
pydantic